Extracted Button and Episode Cover components
