+++ src/components/common/player/components/Button.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { View, Image, TouchableOpacity } from 'react-native';
+
+import styled from 'styled-components';
+import Icon from '../../Icon';
+
+const Touchable = styled(TouchableOpacity)`
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+`;
+
+const Button = ({ onPress, name, size }) => (
+  <Touchable
+    onPress={onPress}
+  >
+    <Icon
+      color="black"
+      name={name}
+      size={size}
+    />
+  </Touchable>
+);
+
+export default Button;
